Change the comparison process to following,
(1) Iterate objects in file 1 and generate a vector of object information.

(2) For each object in file 2, check if there is a object in the vector we created in step 1
is logically equal to it. If no, file 2 is not logically equal to file 1. At the same time,
we generat a vector whose entries are pairs of objects from file 1 and file 2 that are logically
equal to each other.

(3) Strictly/exactly compare pairs of objects, if all pairs are strictly/exactly equal, then
file 1 is strictly/exactly equal to file 2.